requirements.txt
streamlit>=1.49.1
google-genai>=1.32.0
Pillow>=11.3.0
python-dotenv>=1.1.1